refactor code for vr galaxy
